requirements.txt
Flask
requests
spotipy
SpeechRecognition
python-dotenv
PyAudio
